// src/proto/citadel_gcmessages_client.ts
export interface CMsgClientToGCGetMatchHistory {
  /**
   * @generated from protobuf field: optional uint32 account_id = 1;
   */
  accountId?: number;
  /**
   * @generated from protobuf field: optional uint64 continue_cursor = 2;
   */
  continueCursor?: bigint;
}
/**
 * @generated from protobuf message CMsgClientToGCGetMatchHistoryResponse
 */
export interface CMsgClientToGCGetMatchHistoryResponse {
  /**
   * @generated from protobuf field: optional CMsgClientToGCGetMatchHistoryResponse.EResult result = 1;
   */
  result?: CMsgClientToGCGetMatchHistoryResponse_EResult;
  /**
   * @generated from protobuf field: optional uint64 continue_cursor = 2;
   */
  continueCursor?: bigint;
  /**
   * @generated from protobuf field: repeated CMsgClientToGCGetMatchHistoryResponse.Match matches = 3;
   */
  matches: CMsgClientToGCGetMatchHistoryResponse_Match[];
}
/**
 * @generated from protobuf message CMsgClientToGCGetMatchHistoryResponse.Match
 */
export interface CMsgClientToGCGetMatchHistoryResponse_Match {
  /**
   * @generated from protobuf field: optional uint64 match_id = 1;
   */
  matchId?: bigint;
  /**
   * @generated from protobuf field: optional uint32 hero_id = 2;
   */
  heroId?: number;
  /**
   * @generated from protobuf field: optional uint32 match_duration_s = 3;
   */
  matchDurationS?: number;
  /**
   * @generated from protobuf field: optional uint32 start_time = 4;
   */
  startTime?: number;
  /**
   * @generated from protobuf field: optional uint32 match_result = 5;
   */
  matchResult?: number;
  /**
   * @generated from protobuf field: optional ECitadelLobbyTeam player_team = 6;
   */
  playerTeam?: ECitadelLobbyTeam;
  /**
   * @generated from protobuf field: optional uint32 player_kills = 7;
   */
  playerKills?: number;
  /**
   * @generated from protobuf field: optional uint32 player_deaths = 8;
   */
  playerDeaths?: number;
  /**
   * @generated from protobuf field: optional uint32 player_assists = 9;
   */
  playerAssists?: number;
  /**
   * @generated from protobuf field: optional uint32 last_hits = 11;
   */
  lastHits?: number;
  /**
   * @generated from protobuf field: optional uint32 denies = 12;
   */
  denies?: number;
  /**
   * @generated from protobuf field: optional uint32 hero_level = 13;
   */
  heroLevel?: number;
  /**
   * @generated from protobuf field: optional uint32 net_worth = 14;
   */
  netWorth?: number;
  /**
   * @generated from protobuf field: optional uint64 objectives_mask_team0 = 15;
   */
  objectivesMaskTeam0?: bigint;
  /**
   * @generated from protobuf field: optional uint64 objectives_mask_team1 = 16;
   */
  objectivesMaskTeam1?: bigint;
  /**
   * @generated from protobuf field: optional bool team_abandoned = 17;
   */
  teamAbandoned?: boolean;
  /**
   * @generated from protobuf field: optional uint32 abandoned_time_s = 18;
   */
  abandonedTimeS?: number;
}
/**
 * @generated from protobuf enum CMsgClientToGCGetMatchHistoryResponse.EResult
 */
export enum CMsgClientToGCGetMatchHistoryResponse_EResult {
  /**
   * @generated from protobuf enum value: k_eResult_InternalError = 0;
   */
  k_eResult_InternalError = 0,
  /**
   * @generated from protobuf enum value: k_eResult_Success = 1;
   */
  k_eResult_Success = 1,
  /**
   * @generated from protobuf enum value: k_eResult_InvalidPermission = 2;
   */
  k_eResult_InvalidPermission = 2,
  /**
   * @generated from protobuf enum value: k_eResult_TemporarilyDisabled = 3;
   */
  k_eResult_TemporarilyDisabled = 3,
  /**
   * @generated from protobuf enum value: k_eResult_TooBusy = 4;
   */
  k_eResult_TooBusy = 4,
  /**
   * @generated from protobuf enum value: k_eResult_RateLimited = 5;
   */
  k_eResult_RateLimited = 5,
}
// ...
class CMsgClientToGCGetMatchHistory$Type extends MessageType<CMsgClientToGCGetMatchHistory> {
  constructor() {
    super("CMsgClientToGCGetMatchHistory", [
      {
        no: 1,
        name: "account_id",
        kind: "scalar",
        opt: true,
        T: 13 /*ScalarType.UINT32*/,
      },
      {
        no: 2,
        name: "continue_cursor",
        kind: "scalar",
        opt: true,
        T: 4 /*ScalarType.UINT64*/,
        L: 0 /*LongType.BIGINT*/,
      },
    ]);
  }
  create(
    value?: PartialMessage<CMsgClientToGCGetMatchHistory>,
  ): CMsgClientToGCGetMatchHistory {
    const message = globalThis.Object.create(this.messagePrototype!);
    if (value !== undefined)
      reflectionMergePartial<CMsgClientToGCGetMatchHistory>(
        this,
        message,
        value,
      );
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: CMsgClientToGCGetMatchHistory,
  ): CMsgClientToGCGetMatchHistory {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* optional uint32 account_id */ 1:
          message.accountId = reader.uint32();
          break;
        case /* optional uint64 continue_cursor */ 2:
          message.continueCursor = reader.uint64().toBigInt();
          break;
        default:
          let u = options.readUnknownField;
          if (u === "throw")
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: CMsgClientToGCGetMatchHistory,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* optional uint32 account_id = 1; */
    if (message.accountId !== undefined)
      writer.tag(1, WireType.Varint).uint32(message.accountId);
    /* optional uint64 continue_cursor = 2; */
    if (message.continueCursor !== undefined)
      writer.tag(2, WireType.Varint).uint64(message.continueCursor);
    let u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message CMsgClientToGCGetMatchHistory
 */
export const CMsgClientToGCGetMatchHistory =
  new CMsgClientToGCGetMatchHistory$Type();
// @generated message type with reflection information, may provide speed optimized methods
class CMsgClientToGCGetMatchHistoryResponse$Type extends MessageType<CMsgClientToGCGetMatchHistoryResponse> {
  constructor() {
    super("CMsgClientToGCGetMatchHistoryResponse", [
      {
        no: 1,
        name: "result",
        kind: "enum",
        opt: true,
        T: () => [
          "CMsgClientToGCGetMatchHistoryResponse.EResult",
          CMsgClientToGCGetMatchHistoryResponse_EResult,
        ],
      },
      {
        no: 2,
        name: "continue_cursor",
        kind: "scalar",
        opt: true,
        T: 4 /*ScalarType.UINT64*/,
        L: 0 /*LongType.BIGINT*/,
      },
      {
        no: 3,
        name: "matches",
        kind: "message",
        repeat: 2 /*RepeatType.UNPACKED*/,
        T: () => CMsgClientToGCGetMatchHistoryResponse_Match,
      },
    ]);
  }
  create(
    value?: PartialMessage<CMsgClientToGCGetMatchHistoryResponse>,
  ): CMsgClientToGCGetMatchHistoryResponse {
    const message = globalThis.Object.create(this.messagePrototype!);
    message.matches = [];
    if (value !== undefined)
      reflectionMergePartial<CMsgClientToGCGetMatchHistoryResponse>(
        this,
        message,
        value,
      );
    return message;
  }
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: CMsgClientToGCGetMatchHistoryResponse,
  ): CMsgClientToGCGetMatchHistoryResponse {
    let message = target ?? this.create(),
      end = reader.pos + length;
    while (reader.pos < end) {
      let [fieldNo, wireType] = reader.tag();
      switch (fieldNo) {
        case /* optional CMsgClientToGCGetMatchHistoryResponse.EResult result */ 1:
          message.result = reader.int32();
          break;
        case /* optional uint64 continue_cursor */ 2:
          message.continueCursor = reader.uint64().toBigInt();
          break;
        case /* repeated CMsgClientToGCGetMatchHistoryResponse.Match matches */ 3:
          message.matches.push(
            CMsgClientToGCGetMatchHistoryResponse_Match.internalBinaryRead(
              reader,
              reader.uint32(),
              options,
            ),
          );
          break;
        default:
          let u = options.readUnknownField;
          if (u === "throw")
            throw new globalThis.Error(
              `Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`,
            );
          let d = reader.skip(wireType);
          if (u !== false)
            (u === true ? UnknownFieldHandler.onRead : u)(
              this.typeName,
              message,
              fieldNo,
              wireType,
              d,
            );
      }
    }
    return message;
  }
  internalBinaryWrite(
    message: CMsgClientToGCGetMatchHistoryResponse,
    writer: IBinaryWriter,
    options: BinaryWriteOptions,
  ): IBinaryWriter {
    /* optional CMsgClientToGCGetMatchHistoryResponse.EResult result = 1; */
    if (message.result !== undefined)
      writer.tag(1, WireType.Varint).int32(message.result);
    /* optional uint64 continue_cursor = 2; */
    if (message.continueCursor !== undefined)
      writer.tag(2, WireType.Varint).uint64(message.continueCursor);
    /* repeated CMsgClientToGCGetMatchHistoryResponse.Match matches = 3; */
    for (let i = 0; i < message.matches.length; i++)
      CMsgClientToGCGetMatchHistoryResponse_Match.internalBinaryWrite(
        message.matches[i],
        writer.tag(3, WireType.LengthDelimited).fork(),
        options,
      ).join();
    let u = options.writeUnknownFields;
    if (u !== false)
      (u == true ? UnknownFieldHandler.onWrite : u)(
        this.typeName,
        message,
        writer,
      );
    return writer;
  }
}
/**
 * @generated MessageType for protobuf message CMsgClientToGCGetMatchHistoryResponse
 */
export const CMsgClientToGCGetMatchHistoryResponse =
  new CMsgClientToGCGetMatchHistoryResponse$Type();
// ...
export const CMsgClientToGCGetMatchHistoryResponse_Match =
  new CMsgClientToGCGetMatchHistoryResponse_Match$Type();